Tests are not applicable for .NET Framework
